PB-1213: Tidy up code, unify same function.
