+++ src/modules/drawing/components/useExtendLineInteraction.composable.js
@@ -1,13 +1,13 @@
 import { EditableFeatureTypes } from '@/api/features/EditableFeature.class'
 import useDrawingModeInteraction from '@/modules/drawing/components/useDrawingModeInteraction.composable'
 import { drawLineStyle } from '@/modules/drawing/lib/style'
-export default function useContinueLineInteraction({
+
+export default function useExtendLineInteraction({
     style = drawLineStyle,
     featureType = EditableFeatureTypes.LINEPOLYGON,
     drawEndCallback = null,
     startingFeature = null,
 }) {
-    console.log('[useContinueLineInteraction] startingFeature:', startingFeature)
     const { removeLastPoint } = useDrawingModeInteraction({
         geometryType: 'Polygon',
         editingStyle: style,
